+++ robotidy/app.py
@@ -34,6 +34,7 @@ class Robotidy:
         check: bool,
         output: Optional[Path],
         force_order: bool,
+        target_version: int,
     ):
         self.sources = get_paths(src, exclude, extend_exclude)
         self.overwrite = overwrite
@@ -43,7 +44,9 @@ class Robotidy:
         self.formatting_config = formatting_config
         self.output = output
         transformers_config = self.convert_configure(transformers_config)
-        self.transformers = load_transformers(transformers, transformers_config, force_order=force_order)
+        self.transformers = load_transformers(
+            transformers, transformers_config, force_order=force_order, target_version=target_version
+        )
         for transformer in self.transformers:
             # inject global settings TODO: handle it better
             setattr(transformer, "formatting_config", self.formatting_config)